Prediction Model/requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
xgboost==1.7.6
scikit-learn==1.3.0
holidays==0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1